* commit '2e9a1faf6a89d464f4f95dd60008a3ccf4a62b76':
  Tidy x86 assembler.

+++ compiler/dex/quick/arm64/assemble_arm64.cc
@@ -887,7 +887,7 @@ void Arm64Mir2Lir::AssembleLIR() {
   CreateNativeGcMap();
 }
 
-int Arm64Mir2Lir::GetInsnSize(LIR* lir) {
+size_t Arm64Mir2Lir::GetInsnSize(LIR* lir) {
   ArmOpcode opcode = UNWIDE(lir->opcode);
   DCHECK(!IsPseudoLirOp(opcode));
   return EncodingMap[opcode].size;
